requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.13.3
boto3==1.37.13
botocore==1.37.13
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dj-database-url==2.3.0
Django==4.2.20
django-allauth==65.4.1
django-bootstrap-v5==1.0.11
django-bootstrap5==25.1
django-storages==1.14.5
fonttools==4.56.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.68.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.6
psycopg-binary==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
python-barcode==0.15.1 #new